test: consolidate mock client factories into tests/_factories.py
Move write_real_wav / make_stems_client_factory / make_midi_client_factory
into a shared tests/_factories.py and import from both test_export_cli.py and
test_daw_export.py, removing the duplicated definitions (CodeRabbit/claude-review
follow-up). Also untrack tasks/todo.md — it is planning scratch, not repo
content (this time without re-adding it).
